Deleted all previous changes, made this project suitable for creating and showing hotel booking data.
