--- Models/DogViewModel.cs
@@ -1,8 +1,0 @@
-﻿namespace MVCApp.Models
-{
-    public class DogViewModel
-    {
-        public int    Age  { get; set; }
-        public string Name { get; set; }
-    }
-}
