+++ Models/HotelBookingModel.cs
@@ -1,0 +1,10 @@
+﻿namespace MVCApp.Models
+{
+    public class HotelBookingModel
+    {
+        public int      Id        { get; set; }
+        public string   GuestName { get; set; }
+        public DateTime DateStart { get; set; }
+        public DateTime DateEnd   { get; set; }
+    }
+}
