--- Views/HelloWorld/Create.cshtml
@@ -1,31 +1,0 @@
-﻿@model DogViewModel
-
-<h1>Create</h1>
-<hr />
-<div class="row">
-    <div class="col-6">
-        <div class="card">
-            <div class="card-header">
-                <h5>Create a new dog</h5>
-            </div>
-            <div class="card-body">
-                <form asp-controller="HelloWorld" asp-action="CreateDog" method="post">
-                    <div class="mb-3">
-                        <label class="col-form-label">Dog's Name</label>
-                        <input class="form-control" type="text" asp-for="Name" />
-                    </div>
-                    <div class="mb-3">
-                        <label class="col-form-label">Dog's Age</label>
-                        <input class="form-control" type="number" asp-for="Age" />
-                    </div>
-                    <div class="mb-3">
-                        <button class="btn btn-primary" type="submit">Save</button>
-                    </div>
-                </form>
-            </div>
-            <div class="card-footer">
-
-            </div>
-        </div>
-    </div>
-</div>
