--- Views/HelloWorld/Hello.cshtml
@@ -1,6 +1,0 @@
-﻿@model DogViewModel
-
-<h1>Hello World</h1>
-<p>This is the hello page of the HelloWorld Controller</p>
-<h4>A dog named @Model.Name says hello</h4>
-<h4>@Model.Name is @Model.Age years old</h4>
