--- Views/HelloWorld/Index.cshtml
@@ -1,10 +1,0 @@
-﻿@model List<DogViewModel>
-
-<h1>Hello World</h1>
-<p>This is the index page of the HelloWorld Controller</p>
-<ul>
-    @foreach (var dog in Model)
-    {
-        <li>@dog.Name (@dog.Age)</li>
-    }
-</ul>
