+++ Views/HotelBooking/Create.cshtml
@@ -1,0 +1,38 @@
+﻿@model HotelBookingModel
+
+<h1>Create</h1>
+<hr />
+<div class="row">
+    <div class="col-6">
+        <div class="card">
+            <div class="card-header">
+                <h5>Create a new hotel booking</h5>
+            </div>
+            <div class="card-body">
+                <form asp-controller="HotelBooking" asp-action="CreateHotelBooking" method="post">
+                    <div class="mb-3">
+                        <label class="col-form-label">ID:</label>
+                        <input class="form-control" type="number" asp-for="Id" />
+                    </div>
+                    <div class="mb-3">
+                        <label class="col-form-label">Guest's name:</label>
+                        <input class="form-control" type="text" asp-for="GuestName" />
+                    </div>
+                    <div class="mb-3">
+                        <label class="col-form-label">Start date:</label>
+                        <input class="form-control" type="date" asp-for="DateStart" />
+                    </div>
+                    <div class="mb-3">
+                        <label class="col-form-label">Dog's Age</label>
+                        <input class="form-control" type="date" asp-for="DateEnd" />
+                    </div>
+                    <div class="mb-3">
+                        <button class="btn btn-primary" type="submit">Save</button>
+                    </div>
+                </form>
+            </div>
+            <div class="card-footer">
+            </div>
+        </div>
+    </div>
+</div>
