+++ Views/Shared/_Layout.cshtml
@@ -12,7 +12,7 @@
     <header>
         <nav class="navbar navbar-expand-sm navbar-toggleable-sm navbar-light bg-white border-bottom box-shadow mb-3">
             <div class="container-fluid">
-                <a class="navbar-brand" asp-area="" asp-controller="HelloWorld" asp-action="Index">DogApp</a>
+                <a class="navbar-brand" asp-area="" asp-controller="HotelBooking" asp-action="Index">HotelBookingApp</a>
                 <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target=".navbar-collapse" aria-controls="navbarSupportedContent"
                         aria-expanded="false" aria-label="Toggle navigation">
                     <span class="navbar-toggler-icon"></span>
@@ -20,10 +20,10 @@
                 <div class="navbar-collapse collapse d-sm-inline-flex justify-content-between">
                     <ul class="navbar-nav flex-grow-1">
                         <li class="nav-item">
-                            <a class="nav-link text-dark" asp-area="" asp-controller="HelloWorld" asp-action="Index">Dog List</a>
+                            <a class="nav-link text-dark" asp-area="" asp-controller="HotelBooking" asp-action="Index">Hotel Bookings List</a>
                         </li>
                         <li class="nav-item">
-                            <a class="nav-link text-dark" asp-area="" asp-controller="HelloWorld" asp-action="Create">Create</a>
+                            <a class="nav-link text-dark" asp-area="" asp-controller="HotelBooking" asp-action="Create">Create</a>
                         </li>
                     </ul>
                 </div>
